requirements.txt
boto3
botocore
pydantic
python-dotenv
langchain-openai
langchain-core
langchain-community
langgraph
langchain-mcp-adapters
langchain-aws